WIP on separating MetricListItem
