+++ src/Components/MetricList/MetricList.tsx
@@ -1,18 +1,12 @@
 import * as React from "react";
-import { format, fromUnixTime, getUnixTime } from "date-fns";
 import ArrowBoldDownIcon from "@skbkontur/react-icons/ArrowBoldDown";
 import ArrowBoldUpIcon from "@skbkontur/react-icons/ArrowBoldUp";
-import UserIcon from "@skbkontur/react-icons/User";
 import TrashIcon from "@skbkontur/react-icons/Trash";
-import { Tooltip } from "@skbkontur/react-ui/components/Tooltip";
 import { Button } from "@skbkontur/react-ui/components/Button";
 import { MetricItemList } from "../../Domain/Metric";
-import { getUTCDate, humanizeDuration } from "../../helpers/DateUtil";
-import StatusIndicator from "../StatusIndicator/StatusIndicator";
-import MetricValues from "../MetricValues/MetricValues";
-import MaintenanceSelect from "../MaintenanceSelect/MaintenanceSelect";
 import cn from "./MetricList.less";
 import { FixedSizeList as List } from "react-window";
+import { MetricListItem } from "../MetricListItem/MetricListItem";
 
 export type SortingColumn = "state" | "name" | "event" | "value";
 
@@ -28,14 +22,6 @@ type Props = {
     onNoDataRemove?: () => void;
 };
 
-function maintenanceCaption(delta: number): React.ReactNode {
-    return <span>{delta <= 0 ? "Maintenance" : humanizeDuration(delta)}</span>;
-}
-
-function maintenanceDelta(maintenance?: number | null): number {
-    return (maintenance || 0) - getUnixTime(getUTCDate());
-}
-
 export default function MetricList(props: Props): React.ReactElement {
     const {
         status,
@@ -50,10 +36,7 @@ export default function MetricList(props: Props): React.ReactElement {
     } = props;
 
     const sortingIcon = sortingDown ? <ArrowBoldDownIcon /> : <ArrowBoldUpIcon />;
-    const hideTargetsNames = Object.keys(items).every((metric) => {
-        const { values } = items[metric];
-        return !values || Object.keys(values).length === 1;
-    });
+
     const entries = Object.entries(items);
 
     return (
@@ -125,83 +108,16 @@ export default function MetricList(props: Props): React.ReactElement {
                     itemCount={entries.length}
                     itemData={entries}
                 >
-                    {({ data, index, style }) => {
-                        const [metricName, metricData] = data[index];
-                        const {
-                            value,
-                            values,
-                            event_timestamp: eventTimestamp = 0,
-                            state,
-                            maintenance,
-                            maintenance_info: maintenanceInfo,
-                        } = metricData;
-                        const delta = maintenanceDelta(maintenance);
-                        return (
-                            <div key={metricName} className={cn("row")} style={style}>
-                                {status && (
-                                    <div className={cn("state")}>
-                                        <StatusIndicator statuses={[state]} size={10} />
-                                    </div>
-                                )}
-                                <div className={cn("name")}>{metricName}</div>
-                                <div className={cn("event")}>
-                                    {format(fromUnixTime(eventTimestamp), "MMM d, y, HH:mm:ss")}
-                                </div>
-                                <div className={cn("value")}>
-                                    <MetricValues
-                                        value={value}
-                                        values={values}
-                                        placeholder
-                                        hideTargetsNames={hideTargetsNames}
-                                    />
-                                </div>
-                                <div className={cn("maintenance")}>
-                                    <MaintenanceSelect
-                                        maintenance={maintenance}
-                                        caption={maintenanceCaption(delta)}
-                                        onSetMaintenance={(maintenanceValue) =>
-                                            onChange(metricName, maintenanceValue)
-                                        }
-                                    />
-                                </div>
-                                <div className={cn("author")}>
-                                    {delta > 0 &&
-                                        maintenanceInfo &&
-                                        maintenanceInfo.setup_user &&
-                                        maintenanceInfo.setup_time && (
-                                            <Tooltip
-                                                render={() => (
-                                                    <div>
-                                                        Maintenance was set
-                                                        <br />
-                                                        by {maintenanceInfo.setup_user}
-                                                        <br />
-                                                        at{" "}
-                                                        {format(
-                                                            fromUnixTime(
-                                                                maintenanceInfo.setup_time || 0
-                                                            ),
-                                                            "MMMM d, HH:mm:ss"
-                                                        )}
-                                                    </div>
-                                                )}
-                                            >
-                                                <UserIcon className={cn("maintenance-info")} />
-                                            </Tooltip>
-                                        )}
-                                </div>
-                                <div className={cn("delete")}>
-                                    <Button
-                                        use="link"
-                                        icon={<TrashIcon />}
-                                        onClick={() => onRemove(metricName)}
-                                    >
-                                        Delete
-                                    </Button>
-                                </div>
-                            </div>
-                        );
-                    }}
+                    {({ data, index, style }) => (
+                        <MetricListItem
+                            status={status}
+                            data={data}
+                            index={index}
+                            style={style}
+                            onChange={onChange}
+                            onRemove={onRemove}
+                        />
+                    )}
                 </List>
             </div>
         </section>
